Move DAP 2 parsing to dap2-parser subproject
