+++ core/src/test/scala/latis/input/TestGranuleListJoin.scala
@@ -9,9 +9,9 @@ import latis.metadata.Metadata
 import latis.model._
 import latis.ops.GranuleListJoin
 import latis.ops.Selection
-import latis.ops.parser.ast
 import latis.util.Identifier.IdentifierStringContext
 import latis.util.StreamUtils
+import latis.util.dap2.parser.ast
 
 class TestGranuleListJoin {
   //TODO: ScalaTest flat spec?
