+++ dap2-parser/src/main/scala/latis/util/dap2/parser/ConstraintParser.scala
@@ -3,8 +3,8 @@ package latis.util.dap2.parser
 import atto._, Atto._
 import cats.syntax.all._
 
-import latis.ops.parser.ast.CExpr
-import latis.ops.parser.parsers._
+import ast._
+import parsers._
 
 /**
  * Module for parsing DAP 2 constraint expressions.
@@ -32,5 +32,4 @@ object ConstraintParser {
       Right(ast.ConstraintExpression(List.empty))
     }
   }
-
 }
